Add files via upload
This is a simple book library app that I have developed using Android Studio. I have used different libraries such as Volley, Room, Picasso among others for the development of this app. There is a dashboard screen that displays the list of books along with its authors. Upon clicking on each of these books, you are taken to the description of the selected book with a button to add the book to your favorites list at the bottom. There is also a profile page.